+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Hash exclusively using `blake3`
 - change default 'fmode', 'dmode' values
 - rename 'depth', 'width' to 'prefix_depth', 'prefix_width' and change their defaults
+- remove extension preservation on stored blobs
 
 ## v0.8.1